pyproject.toml
[project]
name = "taxonomy-cpt-scraper"
version = "1.0.0"
description = "Taxonomy scraper"
authors = [
    {name = "Ose Dania",email = "o.dania@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.8,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask (>=3.0.0,<4.0.0)"
]

# Disable package mode since we're using this as a script, not a package
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
